Fix namings
Use camelcase for variables
