+++ Faberbabel/Classes/String+Translation.swift
@@ -16,7 +16,7 @@ extension String {
     public func fb_translate(to lang: String) -> String {
         var localized = fb_localize(to: lang)
         if localized == self {
-            EventNotifier.shared?.notify(events: [Event(type: .missing_key, key: self)])
+            EventNotifier.shared?.notify(events: [Event(type: .missingKey, key: self)])
             if lang != "en" {
                 localized = fb_localize(to: "en")
             }
